recognizing minimal keywords
added primitive types keywords
added post-lexer pass

+++ src/lex/token.cpp
@@ -1,10 +1,12 @@
 #include "token.h"
 
+#include <algorithm>
+
 namespace lex {
 
     namespace {
         std::vector<string32> _lex_text_table = {};
-        constexpr auto MAX_TYPE_STR_LEN = 30;
+        constexpr auto MAX_TYPE_STR_LEN = 30ll;
     }
     
     string32& token_t::getText() noexcept {
@@ -17,7 +19,7 @@ namespace lex {
     std::wostream &operator<<(std::wostream &os, const token_stream_t &stream) {
         for (const auto token : stream) {
             const auto& type_cstr = token_to_string_table[(int)token.kind];
-            const auto remaining_space = std::max(0ull, MAX_TYPE_STR_LEN - type_cstr.size());
+            const auto remaining_space = std::max(0ll, MAX_TYPE_STR_LEN - (int64_t)type_cstr.size());
             os << type_cstr;
             if (token.hasText()) {
                 const string32 spaces(remaining_space, L' ');
